os pesos são actually unsigned short, h_free
o h_free() deveria fazer free do item,
mas como o nosso item (int) não é freeable, dá-nos bué chatice.
tirei o free_item do h_free() entretanto.

+++ src/file.h
@@ -5,9 +5,9 @@
 
 #include "graph.h"
 
-int *find_max_perms(FILE *fpal);
+unsigned short *find_max_perms(FILE *fpal);
 
-Graph **read_dic(FILE *fdic, int *max_perms);
+Graph **read_dic(FILE *fdic, unsigned short *max_perms);
 
 void solve_pal(FILE *fpal, FILE *fpath, Graph **graphs);
 void walk_tree(Graph *g, int *st, int *wt, int dst, FILE *);
